+++ src/frontend/apps/accounts/src/pages/index.tsx
@@ -1,7 +1,9 @@
 import { Button } from '@gouvfr-lasuite/cunningham-react';
 import { useTranslation } from 'react-i18next';
 
-import { login, logout, useAuth } from '@/features/auth/Auth';
+import { login, useAuth } from '@/features/auth/Auth';
+import {LOGOUT_URL} from "@/features/auth/conf";
+import {getCSRFToken} from "@/features/api/fetchApi";
 
 export default function HomePage() {
   const { t } = useTranslation();
@@ -12,7 +14,10 @@ export default function HomePage() {
       <main className="welcome">
         <div>
           <h1>{t('Welcome on LaSuite Account')} </h1>
-            <Button onClick={() => logout()}>{t('Logout')}</Button>
+          <form action={LOGOUT_URL} method="post">
+            <input type="hidden" name="csrfmiddlewaretoken" value={getCSRFToken()}/>
+            <Button>{t('Logout')}</Button>
+          </form>
         </div>
       </main>
     );